refactor: extract CBOR scanner into internal module
