+++ webauthn-cbor-internal/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    id("webauthn.kotlin.multiplatform")
+}
+
+kotlin {
+    jvm()
+    iosX64()
+    iosArm64()
+    iosSimulatorArm64()
+
+    sourceSets {
+        commonMain.dependencies {}
+        commonTest.dependencies {
+            implementation(kotlin("test"))
+        }
+    }
+}
