+++ webauthn-server-jvm-crypto/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
     api(project(":webauthn-crypto-api"))
+    implementation(project(":webauthn-cbor-internal"))
     implementation(project(":webauthn-model"))
     implementation(libs.signum.supreme.jvm)
     implementation(libs.signum.indispensable.cosef.jvm)
